requirements.txt
faiss-cpu
numpy==1.26.3
openai==1.8.0
langchain-openai==0.0.3
scikit-learn
sentence-transformers>=2.3.0
tenacity==8.2.3
tiktoken==0.5.2
torch
transformers==4.38.1
umap-learn==0.5.5
urllib3==1.26.6 
setuptools
dotenv
tree_sitter_python
jsonlines
codetext
langchain==0.1.14
datasets
httpx==0.26.0
plotly
python-igraph
google.generativeai
hdbscan